refactor: moving ConfigFormContainer and children into app-config-form sub-directory
