+++ src/pages-and-resources/discussions/app-config-form/index.js
@@ -1,0 +1,1 @@
+export { default } from './ConfigFormContainer';
