+++ src/pages-and-resources/discussions/app-config-form/messages.js
@@ -1,0 +1,10 @@
+import { defineMessages } from '@edx/frontend-platform/i18n';
+
+const messages = defineMessages({
+  configureApp: {
+    id: 'authoring.discussions.configure.app',
+    defaultMessage: 'Configure {name}',
+  },
+});
+
+export default messages;
